recipes-kernel: jailhouse: Enable support for k3-am625-sk platform
Add and enable Jailhouse support for TI's k3-am625-sk similarly to how
other TI platforms were added and enabled in the past by adding platform-
specific configuration variables for the root cell DTB overlay, and the
Linux demo inmate DTB, configuration, and Kernel command line settings.

Furthermore update the platform-specific config files to such that the
root cell DTB gets included in the file system images and that we are
enabling an additional console on ttyS3 (usually mapped to main_uart1)
for the Linux demo inmate to use.

Note that we are also introducing two patches for the Jailhouse source
tree which should get pushed upstream and then dropped here.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>